requirements.txt
aiogram==3.14.0
python-dotenv==1.0.1
Pillow==11.0.0
imageio==2.36.0
fonttools==4.55.3
rlottie-python==1.3.8
